+++ src/renderer/providers/tiktok/components/visualizations/Viz2/data.ts
@@ -20,13 +20,13 @@ const excludeTags = new Set([
 const isValidTag = (tag) => !excludeTags.has(tag);
 
 const transformData = (gdprData, metadata) => {
-  const history = gdprData.Activity["Video Browsing History"].VideoList.map(
-    (d) => ({
-      id: (d.Link ?? d.VideoLink).split("/").reverse()[1],
-      date: new Date(d.Date),
-      ...d,
-    }),
-  );
+  const history = (
+    gdprData.Activity["Video Browsing History"].VideoList ?? []
+  ).map((d) => ({
+    id: (d.Link ?? d.VideoLink).split("/").reverse()[1],
+    date: new Date(d.Date),
+    ...d,
+  }));
 
   const historyMap = group(history, (d) => d.id);